requirements.txt
pillow
cython
matplotlib
tensorboard
tensorboardX
tfrecord
gpustat
tqdm
ipdb
jupyterlab
gdown
hydra-core
clean-fid
wandb
pytorch-lightning
sklearn
hydra-joblib-launcher
hydra_colorlog
python-dotenv
jammy
torchmetrics
einops
jax[cpu]
invisible-watermark
omegaconf
opencv-python==4.5.5.64 
opencv-python-headless==4.5.5.64
diffusers
transformers
kornia